Rename NetCDFInterface as NetCDFPipe
There is an overload of the term interface. Remove move this
by defining the output format (netcdf file) as a pipe, that can
be used to read or write data.

The way the data is passed through this pipe is then labelled the
interface, here the example is VectorInterface.

Restructure the source files accordingly